feat: add browser export
