+++ package.json
@@ -8,6 +8,7 @@
   "type": "module",
   "exports": {
     ".": "./dist/index.js",
+    "./browser": "./dist/browser.js",
     "./helpers": "./dist/helpers/index.js",
     "./package.json": "./package.json"
   },
