+++ src/browser.ts
@@ -1,4 +1,3 @@
-export * from "./code";
 export * from "./core";
 export * from "./types";
 export * from "./format";
