+++ src/core.ts
@@ -1,0 +1,73 @@
+import { print, parse, Options as ParseOptions, types } from "recast";
+import { getBabelParser } from "./babel";
+import {
+  ASTNode,
+  GenerateOptions,
+  ParsedFileNode,
+  Proxified,
+  ProxifiedModule,
+} from "./types";
+import { proxifyModule } from "./proxy/module";
+import { detectCodeFormat } from "./format";
+import { proxify } from "./proxy/proxify";
+
+const b = types.builders;
+
+export function parseModule<Exports extends object = any>(
+  code: string,
+  options?: ParseOptions,
+): ProxifiedModule<Exports> {
+  const node: ParsedFileNode = parse(code, {
+    parser: options?.parser || getBabelParser(),
+    ...options,
+  });
+  return proxifyModule(node, code);
+}
+
+export function parseExpression<T>(
+  code: string,
+  options?: ParseOptions,
+): Proxified<T> {
+  const root: ParsedFileNode = parse("(" + code + ")", {
+    parser: options?.parser || getBabelParser(),
+    ...options,
+  });
+  let body: ASTNode = root.program.body[0];
+  if (body.type === "ExpressionStatement") {
+    body = body.expression;
+  }
+  if (body.extra?.parenthesized) {
+    body.extra.parenthesized = false;
+  }
+
+  const mod = {
+    $ast: root,
+    $code: " " + code + " ",
+    $type: "module",
+  } as any as ProxifiedModule;
+
+  return proxify(body, mod);
+}
+
+export function generateCode(
+  node: { $ast: ASTNode } | ASTNode | ProxifiedModule<any>,
+  options: GenerateOptions = {},
+): { code: string; map?: any } {
+  let ast = (node as Proxified).$ast || node;
+
+  if (ast.type === "FunctionExpression") {
+    ast = b.expressionStatement(ast);
+  }
+
+  const formatOptions =
+    options.format === false || !("$code" in node)
+      ? {}
+      : detectCodeFormat(node.$code, options.format);
+
+  const { code, map } = print(ast, {
+    ...options,
+    ...formatOptions,
+  });
+
+  return { code, map };
+}
